5/java-bridge-main/src/main/java/bridge/BridgeGame.java
package bridge;

import java.util.ArrayList;
import java.util.Collections;
import java.util.List;

/**
 * 다리 건너기 게임을 관리하는 클래스
 */
public class BridgeGame {


    private final List<List<String>> result;

    public BridgeGame() {
        result = new ArrayList<>();
    }

    /**
     * 사용자가 칸을 이동할 때 사용하는 메서드
     * <p>
     * 이동을 위해 필요한 메서드의 반환 타입(return type), 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
     */
    public void move(List<String> bridge, String direction) {
        if (bridge.get(result.size()).equals(direction)) {
            result.add(List.of(direction, "O"));
        }

        if (!bridge.get(result.size()).equals(direction)) {
            result.add(List.of(direction, "X"));
        }
    }

    /**
     * 사용자가 게임을 다시 시도할 때 사용하는 메서드
     * <p>
     * 재시작을 위해 필요한 메서드의 반환 타입(return type), 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
     */
    public void retry() {
    }

    public List<List<String>> getResult() {
        return Collections.unmodifiableList(result);
    }
}
